requirements.txt
aiohttp==3.6.2
asn1crypto==1.3.0
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.11.8
cffi==1.14.0
chardet==3.0.4
colorama==0.4.4
cryptography==2.8
discord.py==1.3.3
idna==2.9
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.5
pycparser==2.20
pylint==2.6.0
pyOpenSSL==19.1.0
PySocks==1.7.1
python-dotenv==0.12.0
requests==2.23.0
six==1.14.0
toml==0.10.1
urllib3==1.25.8
websockets==8.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.2
yarl==1.4.2